requirements.txt
Django==1.8.2
Pillow==2.8.2
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.20
celery==3.1.18
django-celery==3.1.16
django-ckeditor==4.4.8
django-widget-tweaks==1.3
elasticsearch==1.6.0
kombu==3.0.26
pika==0.9.14
psycopg2==2.6
pytz==2015.4
urllib3==1.10.4
wsgiref==0.1.2
